@@ -1,5 +1,6 @@
 FactoryGirl.define do
   factory :role do |role|
-    role.sequence(:name) { |n| Faker::Company.position.split(" ")[-1].concat(" #{n}") }
+    rol.name { Faker::Company.position.split(" ")[-1] }
+    # role.sequence(:name) { |n| Faker::Company.position.split(" ")[-1].concat(" #{n}") }
   end
 end
