@@ -2,8 +2,10 @@
 
 FactoryGirl.define do
   factory :user do |user|
-    user.sequence(:name) { |n| Faker::Name.name + " #{n}" }
-    user.sequence(:email) { |n| "#{n}" + Faker::Internet.email }
+    user.name { Faker::Name.name }
+    # user.sequence(:name) { |n| Faker::Name.name + " #{n}" }
+    user.email { Faker::Internet.email }
+    # user.sequence(:email) { |n| "#{n}" + Faker::Internet.email }
     user.password "test_pwd"
     user.password_confirmation "test_pwd"
     user.category {["patient", "hospital_staff", "guest"].sample}
